+++ src/components/molecules/MarkList.tsx
@@ -1,0 +1,44 @@
+import React from 'react';
+import { View, Text, StyleSheet } from 'react-native';
+import ModalIconButton from '@components/atoms/ModalIconButton';
+import { TMarkData } from '@/types/TApiType';
+
+type MarkListProps = {
+  data: TMarkData[];
+  onSelect: (item: TMarkData) => void;
+};
+
+const MarkList = ({ data, onSelect }: MarkListProps) => {
+  if (!data || data.length === 0) {
+    return (
+      <View style={{ flex: 1, alignItems: 'center', marginTop: 40 }}>
+        <Text style={{ color: '#999', fontSize: 16 }}>검색 결과 없음</Text>
+      </View>
+    );
+  }
+  
+  return (
+    <View style={styles.markList}>
+      {data.map((item, idx) => (
+        <ModalIconButton
+          key={`${item.code}-${idx}`}
+          item={item}
+          markSelected={onSelect}
+        />
+      ))}
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  markList: {
+    paddingHorizontal: 2,
+    paddingVertical: '5%',
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
+
+export default MarkList;
