+++ src/components/molecules/SearchInputAndButton.tsx
@@ -1,25 +1,34 @@
 import SearchIdInput from '@components/organisms/SearchIdInput';
 import Button from '@components/atoms/Button';
-import { StyleSheet, Text, View } from 'react-native';
+import { ColorValue, StyleSheet, View } from 'react-native';
 import SearchSvg from '@assets/svgs/search.svg';
 import { font, os } from '@/style/font';
+
+interface ISearchInputAndButtonProps {
+  textInputsObject: {
+    placeholder?: string;
+    placeholderTextColor?: ColorValue;
+    onChangeText: (val: string) => void;
+    value?: string;
+  };
+  buttonClickHandler: () => void;
+}
 
 const SearchInputAndButton = ({
   textInputsObject,
   buttonClickHandler,
-}: any) => {
+}: ISearchInputAndButtonProps) => {
   return (
     <View style={styles.modalSearchWrapper}>
       {/* input이 가로 공간을 최대한 차지하도록 flex:1 추가 */}
-      <View style={{ flex: 1, marginLeft: -10 }}>
+      <View style={{ flex: 1 }}>
         <SearchIdInput textInputs={[textInputsObject]} errorState={false} />
       </View>
 
       {/* 버튼은 내용에 맞는 크기만 차지 */}
       <Button.scale onPress={buttonClickHandler}>
         <View style={styles.searchButton}>
-          <SearchSvg width={16} height={16} color={'#fff'} />
-          <Text style={styles.buttonText}>검색</Text>
+          <SearchSvg width={25} height={16} color={'#fff'} />
         </View>
       </Button.scale>
     </View>
